refactor(multisig): split signature gas helpers
Addresses https://github.com/tempoxyz/tempo/pull/5178#discussion_r3513869565
